+++ assets/js/phoenix_live_view/dom_patch.js
@@ -8,7 +8,8 @@ import {
   PHX_SKIP,
   PHX_STATIC,
   PHX_TRIGGER_ACTION,
-  PHX_UPDATE
+  PHX_UPDATE,
+  PHX_STREAM,
 } from "./constants"
 
 import {
@@ -33,15 +34,19 @@ export default class DOMPatch {
     })
   }
 
-  constructor(view, container, id, html, targetCID){
+  constructor(view, container, id, html, streams, targetCID){
     this.view = view
     this.liveSocket = view.liveSocket
     this.container = container
     this.id = id
     this.rootID = view.root.id
     this.html = html
+    this.streams = streams
+    this.streamInserts = {}
     this.targetCID = targetCID
     this.cidPatch = isCid(this.targetCID)
+    this.pendingRemoves = []
+    this.phxRemove = this.liveSocket.binding("remove")
     this.callbacks = {
       beforeadded: [], beforeupdated: [], beforephxChildAdded: [],
       afteradded: [], afterupdated: [], afterdiscarded: [], afterphxChildAdded: [],
@@ -61,7 +66,9 @@ export default class DOMPatch {
   }
 
   markPrunableContentForRemoval(){
-    DOM.all(this.container, "[phx-update=append] > *, [phx-update=prepend] > *", el => {
+    let phxUpdate = this.liveSocket.binding(PHX_UPDATE)
+    DOM.all(this.container, `[${phxUpdate}=${PHX_STREAM}]`, el => el.innerHTML = "")
+    DOM.all(this.container, `[${phxUpdate}=append] > *, [${phxUpdate}=prepend] > *`, el => {
       el.setAttribute(PHX_PRUNE, "")
     })
   }
@@ -77,11 +84,10 @@ export default class DOMPatch {
     let phxFeedbackFor = liveSocket.binding(PHX_FEEDBACK_FOR)
     let disableWith = liveSocket.binding(PHX_DISABLE_WITH)
     let phxTriggerExternal = liveSocket.binding(PHX_TRIGGER_ACTION)
-    let phxRemove = liveSocket.binding("remove")
     let added = []
     let updates = []
     let appendPrependUpdates = []
-    let pendingRemoves = []
+
     let externalFormTriggered = null
 
     let diffHTML = liveSocket.time("premorph container prep", () => {
@@ -92,10 +98,42 @@ export default class DOMPatch {
     this.trackBefore("updated", container, container)
 
     liveSocket.time("morphdom", () => {
+
+      this.streams.forEach(([inserts, deleteIds]) => {
+        this.streamInserts = Object.assign(this.streamInserts, inserts)
+        deleteIds.forEach(id => {
+          let child = container.querySelector(`[id="${id}"]`)
+          if(child){
+            if(!this.maybePendingRemove(child)){
+              child.remove()
+              this.onNodeDiscarded(child)
+            }
+          }
+        })
+      })
+
       morphdom(targetContainer, diffHTML, {
         childrenOnly: targetContainer.getAttribute(PHX_COMPONENT) === null,
         getNodeKey: (node) => {
           return DOM.isPhxDestroyed(node) ? null : node.id
+        },
+        // skip indexing from children when container is stream
+        skipFromChildren: (from) => { return from.getAttribute(phxUpdate) === PHX_STREAM },
+        // tell morphdom how to add a child
+        addChild: (parent, child) => {
+          let streamAt = child.id ? this.streamInserts[child.id] : undefined
+          if(streamAt === undefined) { return parent.appendChild(child) }
+
+          //streaming
+          DOM.putPrivate(child, PHX_STREAM, true)
+          if(streamAt === 0){
+            parent.insertAdjacentElement("afterbegin", child)
+          } else if(streamAt === -1){
+            parent.appendChild(child)
+          } else if(streamAt > 0){
+            let sibling = Array.from(parent.children)[streamAt]
+            parent.insertBefore(child, sibling)
+          }
         },
         onBeforeNodeAdded: (el) => {
           this.trackBefore("added", el)
@@ -119,18 +157,12 @@ export default class DOMPatch {
           }
           added.push(el)
         },
-        onNodeDiscarded: (el) => {
-          // nested view handling
-          if(DOM.isPhxChild(el) || DOM.isPhxSticky(el)){ liveSocket.destroyViewByEl(el) }
-          this.trackAfter("discarded", el)
-        },
+        onNodeDiscarded: (el) => this.onNodeDiscarded(el),
         onBeforeNodeDiscarded: (el) => {
           if(el.getAttribute && el.getAttribute(PHX_PRUNE) !== null){ return true }
-          if(el.parentNode !== null && DOM.isPhxUpdate(el.parentNode, phxUpdate, ["append", "prepend"]) && el.id){ return false }
-          if(el.getAttribute && el.getAttribute(phxRemove)){
-            pendingRemoves.push(el)
-            return false
-          }
+          if(DOM.private(el, PHX_STREAM)){ return false }
+          if(el.parentElement !== null && DOM.isPhxUpdate(el.parentElement, phxUpdate, ["append", "prepend"]) && el.id){ return false }
+          if(this.maybePendingRemove(el)){ return false }
           if(this.skipCIDSibling(el)){ return false }
           return true
         },
@@ -139,6 +171,7 @@ export default class DOMPatch {
             externalFormTriggered = el
           }
           updates.push(el)
+          this.maybeReOrderStream(el)
         },
         onBeforeElUpdated: (fromEl, toEl) => {
           DOM.cleanChildNodes(toEl, phxUpdate)
@@ -209,6 +242,55 @@ export default class DOMPatch {
     added.forEach(el => this.trackAfter("added", el))
     updates.forEach(el => this.trackAfter("updated", el))
 
+    this.transitionPendingRemoves()
+
+    if(externalFormTriggered){
+      liveSocket.unload()
+      externalFormTriggered.submit()
+    }
+    return true
+  }
+
+  onNodeDiscarded(el){
+    // nested view handling
+    if(DOM.isPhxChild(el) || DOM.isPhxSticky(el)){ this.liveSocket.destroyViewByEl(el) }
+    this.trackAfter("discarded", el)
+  }
+
+  maybePendingRemove(node){
+    if(node.getAttribute && node.getAttribute(this.phxRemove) !== null){
+      this.pendingRemoves.push(node)
+      return true
+    } else {
+      return false
+    }
+  }
+
+  maybeReOrderStream(el){
+    let streamAt = el.id ? this.streamInserts[el.id] : undefined
+    if(streamAt === undefined){ return }
+
+    DOM.putPrivate(el, PHX_STREAM, true)
+    if(streamAt === 0){
+      el.parentElement.insertBefore(el, el.parentElement.firstElementChild)
+    } else if(streamAt > 0){
+      let children = Array.from(el.parentElement.children)
+      let oldIndex = children.indexOf(el)
+      if(streamAt >= children.length - 1){
+        el.parentElement.appendChild(el)
+      } else {
+        let sibling = children[streamAt]
+        if(oldIndex > streamAt){
+          el.parentElement.insertBefore(el, sibling)
+        } else {
+          el.parentElement.insertBefore(el, sibling.nextElementSibling)
+        }
+      }
+    }
+  }
+
+  transitionPendingRemoves(){
+    let {pendingRemoves, liveSocket} = this
     if(pendingRemoves.length > 0){
       liveSocket.transitionRemoves(pendingRemoves)
       liveSocket.requestDOMUpdate(() => {
@@ -220,12 +302,6 @@ export default class DOMPatch {
         this.trackAfter("transitionsDiscarded", pendingRemoves)
       })
     }
-
-    if(externalFormTriggered){
-      liveSocket.unload()
-      externalFormTriggered.submit()
-    }
-    return true
   }
 
   isCIDPatch(){ return this.cidPatch }
@@ -275,4 +351,6 @@ export default class DOMPatch {
       return diffContainer.outerHTML
     }
   }
+
+  indexOf(parent, child){ return Array.from(parent.children).indexOf(child) }
 }