+++ lib/phoenix_live_view/lifecycle.ex
@@ -10,10 +10,11 @@ defmodule Phoenix.LiveView.Lifecycle do
           handle_event: [hook],
           handle_info: [hook],
           handle_params: [hook],
+          after_render: [hook],
           mount: [hook]
         }
 
-  defstruct handle_event: [], handle_info: [], handle_params: [], mount: []
+  defstruct handle_event: [], handle_info: [], handle_params: [], mount: [], after_render: []
 
   @doc """
   Returns a map of infos about the lifecycle stage for the given `view`.
@@ -40,7 +41,7 @@ defmodule Phoenix.LiveView.Lifecycle do
   end
 
   def attach_hook(%Socket{} = socket, id, stage, fun)
-      when stage in [:handle_event, :handle_info, :handle_params] do
+      when stage in [:handle_event, :handle_info, :handle_params, :after_render] do
     lifecycle = lifecycle(socket)
     hook = hook!(id, stage, fun)
     existing = Enum.find(Map.fetch!(lifecycle, stage), &(&1.id == id))
@@ -67,7 +68,7 @@ defmodule Phoenix.LiveView.Lifecycle do
   end
 
   def detach_hook(%Socket{} = socket, id, stage)
-      when stage in [:handle_event, :handle_info, :handle_params] do
+      when stage in [:handle_event, :handle_info, :handle_params, :after_render] do
     update_lifecycle(socket, stage, fn hooks ->
       for hook <- hooks, hook.id != id, do: hook
     end)
@@ -188,6 +189,26 @@ defmodule Phoenix.LiveView.Lifecycle do
     reduce_socket(lifecycle.handle_info, socket, fn hook, acc ->
       hook.function.(msg, acc)
     end)
+  end
+
+  @doc false
+  def after_render(%Socket{private: %{@lifecycle => lifecycle}} = socket) do
+    {:cont, new_socket} =
+      reduce_socket(lifecycle.after_render, socket, fn hook, acc ->
+        case hook.function.(acc) do
+          ^socket ->
+            {:cont, socket}
+
+          %Socket{} = new_socket ->
+            {:cont, Utils.clear_changed(new_socket)}
+
+          other ->
+            raise ArgumentError,
+                  "expected after_render hook to return a socket, got: #{inspect(other)}"
+        end
+      end)
+
+    new_socket
   end
 
   defp reduce_socket([hook | hooks], acc, function) do