+++ priv/static/phoenix_live_view.js
@@ -84,6 +84,7 @@ var LiveView = (() => {
   var PHX_DEBOUNCE = "debounce";
   var PHX_THROTTLE = "throttle";
   var PHX_UPDATE = "update";
+  var PHX_STREAM = "stream";
   var PHX_KEY = "key";
   var PHX_PRIVATE = "phxPrivate";
   var PHX_AUTO_RECOVER = "auto-recover";
@@ -110,6 +111,7 @@ var LiveView = (() => {
   var REPLY = "r";
   var TITLE = "t";
   var TEMPLATES = "p";
+  var STREAM = "stream";
 
   // js/phoenix_live_view/entry_uploader.js
   var EntryUploader = class {
@@ -1297,6 +1299,8 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
         } else {
           toNode = toElement(toNode);
         }
+      } else if (toNode.nodeType === DOCUMENT_FRAGMENT_NODE$1) {
+        toNode = toNode.firstElementChild;
       }
       var getNodeKey = options.getNodeKey || defaultGetNodeKey;
       var onBeforeNodeAdded = options.onBeforeNodeAdded || noop;
@@ -1306,6 +1310,10 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       var onBeforeNodeDiscarded = options.onBeforeNodeDiscarded || noop;
       var onNodeDiscarded = options.onNodeDiscarded || noop;
       var onBeforeElChildrenUpdated = options.onBeforeElChildrenUpdated || noop;
+      var skipFromChildren = options.skipFromChildren || noop;
+      var addChild = options.addChild || function(parent, child) {
+        return parent.appendChild(child);
+      };
       var childrenOnly = options.childrenOnly === true;
       var fromNodesLookup = Object.create(null);
       var keyedRemovalList = [];
@@ -1406,6 +1414,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
         }
       }
       function morphChildren(fromEl, toEl) {
+        var skipFrom = skipFromChildren(fromEl);
         var curToNodeChild = toEl.firstChild;
         var curFromNodeChild = fromEl.firstChild;
         var curToNodeKey;
@@ -1417,7 +1426,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           while (curToNodeChild) {
             toNextSibling = curToNodeChild.nextSibling;
             curToNodeKey = getNodeKey(curToNodeChild);
-            while (curFromNodeChild) {
+            while (!skipFrom && curFromNodeChild) {
               fromNextSibling = curFromNodeChild.nextSibling;
               if (curToNodeChild.isSameNode && curToNodeChild.isSameNode(curFromNodeChild)) {
                 curToNodeChild = toNextSibling;
@@ -1474,7 +1483,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
               curFromNodeChild = fromNextSibling;
             }
             if (curToNodeKey && (matchingFromEl = fromNodesLookup[curToNodeKey]) && compareNodeNames(matchingFromEl, curToNodeChild)) {
-              fromEl.appendChild(matchingFromEl);
+              if (!skipFrom) {
+                addChild(fromEl, matchingFromEl);
+              }
               morphEl(matchingFromEl, curToNodeChild);
             } else {
               var onBeforeNodeAddedResult = onBeforeNodeAdded(curToNodeChild);
@@ -1485,7 +1496,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
                 if (curToNodeChild.actualize) {
                   curToNodeChild = curToNodeChild.actualize(fromEl.ownerDocument || doc);
                 }
-                fromEl.appendChild(curToNodeChild);
+                addChild(fromEl, curToNodeChild);
                 handleNodeAdded(curToNodeChild);
               }
             }
@@ -1563,15 +1574,19 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
         }
       });
     }
-    constructor(view, container, id, html, targetCID) {
+    constructor(view, container, id, html, streams, targetCID) {
       this.view = view;
       this.liveSocket = view.liveSocket;
       this.container = container;
       this.id = id;
       this.rootID = view.root.id;
       this.html = html;
+      this.streams = streams;
+      this.streamInserts = {};
       this.targetCID = targetCID;
       this.cidPatch = isCid(this.targetCID);
+      this.pendingRemoves = [];
+      this.phxRemove = this.liveSocket.binding("remove");
       this.callbacks = {
         beforeadded: [],
         beforeupdated: [],
@@ -1596,7 +1611,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       this.callbacks[`after${kind}`].forEach((callback) => callback(...args));
     }
     markPrunableContentForRemoval() {
-      dom_default.all(this.container, "[phx-update=append] > *, [phx-update=prepend] > *", (el) => {
+      let phxUpdate = this.liveSocket.binding(PHX_UPDATE);
+      dom_default.all(this.container, `[${phxUpdate}=${PHX_STREAM}]`, (el) => el.innerHTML = "");
+      dom_default.all(this.container, `[${phxUpdate}=append] > *, [${phxUpdate}=prepend] > *`, (el) => {
         el.setAttribute(PHX_PRUNE, "");
       });
     }
@@ -1612,11 +1629,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       let phxFeedbackFor = liveSocket.binding(PHX_FEEDBACK_FOR);
       let disableWith = liveSocket.binding(PHX_DISABLE_WITH);
       let phxTriggerExternal = liveSocket.binding(PHX_TRIGGER_ACTION);
-      let phxRemove = liveSocket.binding("remove");
       let added = [];
       let updates = [];
       let appendPrependUpdates = [];
-      let pendingRemoves = [];
       let externalFormTriggered = null;
       let diffHTML = liveSocket.time("premorph container prep", () => {
         return this.buildDiffHTML(container, html, phxUpdate, targetContainer);
@@ -1624,10 +1639,40 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       this.trackBefore("added", container);
       this.trackBefore("updated", container, container);
       liveSocket.time("morphdom", () => {
+        this.streams.forEach(([inserts, deleteIds]) => {
+          this.streamInserts = Object.assign(this.streamInserts, inserts);
+          deleteIds.forEach((id) => {
+            let child = container.querySelector(`[id="${id}"]`);
+            if (child) {
+              if (!this.maybePendingRemove(child)) {
+                child.remove();
+                this.onNodeDiscarded(child);
+              }
+            }
+          });
+        });
         morphdom_esm_default(targetContainer, diffHTML, {
           childrenOnly: targetContainer.getAttribute(PHX_COMPONENT) === null,
           getNodeKey: (node) => {
             return dom_default.isPhxDestroyed(node) ? null : node.id;
+          },
+          skipFromChildren: (from) => {
+            return from.getAttribute(phxUpdate) === PHX_STREAM;
+          },
+          addChild: (parent, child) => {
+            let streamAt = child.id ? this.streamInserts[child.id] : void 0;
+            if (streamAt === void 0) {
+              return parent.appendChild(child);
+            }
+            dom_default.putPrivate(child, PHX_STREAM, true);
+            if (streamAt === 0) {
+              parent.insertAdjacentElement("afterbegin", child);
+            } else if (streamAt === -1) {
+              parent.appendChild(child);
+            } else if (streamAt > 0) {
+              let sibling = Array.from(parent.children)[streamAt];
+              parent.insertBefore(child, sibling);
+            }
           },
           onBeforeNodeAdded: (el) => {
             this.trackBefore("added", el);
@@ -1648,21 +1693,18 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
             }
             added.push(el);
           },
-          onNodeDiscarded: (el) => {
-            if (dom_default.isPhxChild(el) || dom_default.isPhxSticky(el)) {
-              liveSocket.destroyViewByEl(el);
-            }
-            this.trackAfter("discarded", el);
-          },
+          onNodeDiscarded: (el) => this.onNodeDiscarded(el),
           onBeforeNodeDiscarded: (el) => {
             if (el.getAttribute && el.getAttribute(PHX_PRUNE) !== null) {
               return true;
             }
-            if (el.parentNode !== null && dom_default.isPhxUpdate(el.parentNode, phxUpdate, ["append", "prepend"]) && el.id) {
+            if (dom_default.private(el, PHX_STREAM)) {
               return false;
             }
-            if (el.getAttribute && el.getAttribute(phxRemove)) {
-              pendingRemoves.push(el);
+            if (el.parentElement !== null && dom_default.isPhxUpdate(el.parentElement, phxUpdate, ["append", "prepend"]) && el.id) {
+              return false;
+            }
+            if (this.maybePendingRemove(el)) {
               return false;
             }
             if (this.skipCIDSibling(el)) {
@@ -1675,6 +1717,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
               externalFormTriggered = el;
             }
             updates.push(el);
+            this.maybeReOrderStream(el);
           },
           onBeforeElUpdated: (fromEl, toEl) => {
             dom_default.cleanChildNodes(toEl, phxUpdate);
@@ -1746,6 +1789,52 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       dom_default.dispatchEvent(document, "phx:update");
       added.forEach((el) => this.trackAfter("added", el));
       updates.forEach((el) => this.trackAfter("updated", el));
+      this.transitionPendingRemoves();
+      if (externalFormTriggered) {
+        liveSocket.unload();
+        externalFormTriggered.submit();
+      }
+      return true;
+    }
+    onNodeDiscarded(el) {
+      if (dom_default.isPhxChild(el) || dom_default.isPhxSticky(el)) {
+        this.liveSocket.destroyViewByEl(el);
+      }
+      this.trackAfter("discarded", el);
+    }
+    maybePendingRemove(node) {
+      if (node.getAttribute && node.getAttribute(this.phxRemove) !== null) {
+        this.pendingRemoves.push(node);
+        return true;
+      } else {
+        return false;
+      }
+    }
+    maybeReOrderStream(el) {
+      let streamAt = el.id ? this.streamInserts[el.id] : void 0;
+      if (streamAt === void 0) {
+        return;
+      }
+      dom_default.putPrivate(el, PHX_STREAM, true);
+      if (streamAt === 0) {
+        el.parentElement.insertBefore(el, el.parentElement.firstElementChild);
+      } else if (streamAt > 0) {
+        let children = Array.from(el.parentElement.children);
+        let oldIndex = children.indexOf(el);
+        if (streamAt >= children.length - 1) {
+          el.parentElement.appendChild(el);
+        } else {
+          let sibling = children[streamAt];
+          if (oldIndex > streamAt) {
+            el.parentElement.insertBefore(el, sibling);
+          } else {
+            el.parentElement.insertBefore(el, sibling.nextElementSibling);
+          }
+        }
+      }
+    }
+    transitionPendingRemoves() {
+      let { pendingRemoves, liveSocket } = this;
       if (pendingRemoves.length > 0) {
         liveSocket.transitionRemoves(pendingRemoves);
         liveSocket.requestDOMUpdate(() => {
@@ -1759,11 +1848,6 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           this.trackAfter("transitionsDiscarded", pendingRemoves);
         });
       }
-      if (externalFormTriggered) {
-        liveSocket.unload();
-        externalFormTriggered.submit();
-      }
-      return true;
     }
     isCIDPatch() {
       return this.cidPatch;
@@ -1805,6 +1889,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
         return diffContainer.outerHTML;
       }
     }
+    indexOf(parent, child) {
+      return Array.from(parent.children).indexOf(child);
+    }
   };
 
   // js/phoenix_live_view/rendered.js
@@ -1825,13 +1912,14 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       return this.viewId;
     }
     toString(onlyCids) {
-      return this.recursiveToString(this.rendered, this.rendered[COMPONENTS], onlyCids);
+      let [str, streams] = this.recursiveToString(this.rendered, this.rendered[COMPONENTS], onlyCids);
+      return [str, streams];
     }
     recursiveToString(rendered, components = rendered[COMPONENTS], onlyCids) {
       onlyCids = onlyCids ? new Set(onlyCids) : null;
-      let output = { buffer: "", components, onlyCids };
+      let output = { buffer: "", components, onlyCids, streams: new Set() };
       this.toOutputBuffer(rendered, null, output);
-      return output.buffer;
+      return [output.buffer, output.streams];
     }
     componentCIDs(diff) {
       return Object.keys(diff[COMPONENTS] || {}).map((i) => parseInt(i));
@@ -1896,7 +1984,8 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       for (let key in source) {
         let val = source[key];
         let targetVal = target[key];
-        if (isObject(val) && val[STATIC] === void 0 && isObject(targetVal)) {
+        let isObjVal = isObject(val);
+        if (isObjVal && val[STATIC] === void 0 && isObject(targetVal)) {
           this.doMutableMerge(targetVal, val);
         } else {
           target[key] = val;
@@ -1915,7 +2004,8 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       return merged;
     }
     componentToString(cid) {
-      return this.recursiveCIDToString(this.rendered[COMPONENTS], cid);
+      let [str, streams] = this.recursiveCIDToString(this.rendered[COMPONENTS], cid);
+      return [str, streams];
     }
     pruneCIDs(cids) {
       cids.forEach((cid) => delete this.rendered[COMPONENTS][cid]);
@@ -1946,7 +2036,8 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       }
     }
     comprehensionToBuffer(rendered, templates, output) {
-      let { [DYNAMICS]: dynamics, [STATIC]: statics } = rendered;
+      let { [DYNAMICS]: dynamics, [STATIC]: statics, [STREAM]: stream } = rendered;
+      let [_inserts, deleteIds] = stream || [{}, []];
       statics = this.templateStatic(statics, templates);
       let compTemplates = templates || rendered[TEMPLATES];
       for (let d = 0; d < dynamics.length; d++) {
@@ -1957,10 +2048,16 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           output.buffer += statics[i];
         }
       }
+      if (stream !== void 0 && (rendered[DYNAMICS].length > 0 || deleteIds.length > 0)) {
+        rendered[DYNAMICS] = [];
+        output.streams.add(stream);
+      }
     }
     dynamicToBuffer(rendered, templates, output) {
       if (typeof rendered === "number") {
-        output.buffer += this.recursiveCIDToString(output.components, rendered, output.onlyCids);
+        let [str, streams] = this.recursiveCIDToString(output.components, rendered, output.onlyCids);
+        output.buffer += str;
+        output.streams = new Set([...output.streams, ...streams]);
       } else if (isObject(rendered)) {
         this.toOutputBuffer(rendered, templates, output);
       } else {
@@ -1970,7 +2067,8 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
     recursiveCIDToString(components, cid, onlyCids) {
       let component = components[cid] || logError(`no component for CID ${cid}`, components);
       let template = document.createElement("template");
-      template.innerHTML = this.recursiveToString(component, components, onlyCids);
+      let [html, streams] = this.recursiveToString(component, components, onlyCids);
+      template.innerHTML = html;
       let container = template.content;
       let skip = onlyCids && !onlyCids.has(cid);
       let [hasChildNodes, hasChildComponents] = Array.from(container.childNodes).reduce(([hasNodes, hasComponents], child, i) => {
@@ -2005,12 +2103,12 @@ within:
       }, [false, false]);
       if (!hasChildNodes && !hasChildComponents) {
         logError("expected at least one HTML element tag inside a component, but the component is empty:\n", template.innerHTML.trim());
-        return this.createSpan("", cid).outerHTML;
+        return [this.createSpan("", cid).outerHTML, streams];
       } else if (!hasChildNodes && hasChildComponents) {
         logError("expected at least one HTML element tag directly inside a component, but only subcomponents were found. A component must render at least one HTML tag directly inside itself.", template.innerHTML.trim());
-        return template.innerHTML;
+        return [template.innerHTML, streams];
       } else {
-        return template.innerHTML;
+        return [template.innerHTML, streams];
       }
     }
     createSpan(text, cid) {
@@ -2482,7 +2580,7 @@ within:
       browser_default.dropLocal(this.liveSocket.localStorage, window.location.pathname, CONSECUTIVE_RELOADS);
       this.applyDiff("mount", rendered, ({ diff, events }) => {
         this.rendered = new Rendered(this.id, diff);
-        let html = this.renderContainer(null, "join");
+        let [html, streams] = this.renderContainer(null, "join");
         this.dropPendingRefs();
         let forms = this.formsForRecovery(html);
         this.joinCount++;
@@ -2490,12 +2588,12 @@ within:
           forms.forEach(([form, newForm, newCid], i) => {
             this.pushFormRecovery(form, newCid, (resp2) => {
               if (i === forms.length - 1) {
-                this.onJoinComplete(resp2, html, events);
+                this.onJoinComplete(resp2, html, streams, events);
               }
             });
           });
         } else {
-          this.onJoinComplete(resp, html, events);
+          this.onJoinComplete(resp, html, streams, events);
         }
       });
     }
@@ -2505,9 +2603,9 @@ within:
         el.removeAttribute(PHX_REF_SRC);
       });
     }
-    onJoinComplete({ live_patch }, html, events) {
+    onJoinComplete({ live_patch }, html, streams, events) {
       if (this.joinCount > 1 || this.parent && !this.parent.isJoinPending()) {
-        return this.applyJoinPatch(live_patch, html, events);
+        return this.applyJoinPatch(live_patch, html, streams, events);
       }
       let newChildren = dom_default.findPhxChildrenInFragment(html, this.id).filter((toEl) => {
         let fromEl = toEl.id && this.el.querySelector(`[id="${toEl.id}"]`);
@@ -2519,14 +2617,14 @@ within:
       });
       if (newChildren.length === 0) {
         if (this.parent) {
-          this.root.pendingJoinOps.push([this, () => this.applyJoinPatch(live_patch, html, events)]);
+          this.root.pendingJoinOps.push([this, () => this.applyJoinPatch(live_patch, html, streams, events)]);
           this.parent.ackJoin(this);
         } else {
           this.onAllChildJoinsComplete();
-          this.applyJoinPatch(live_patch, html, events);
+          this.applyJoinPatch(live_patch, html, streams, events);
         }
       } else {
-        this.root.pendingJoinOps.push([this, () => this.applyJoinPatch(live_patch, html, events)]);
+        this.root.pendingJoinOps.push([this, () => this.applyJoinPatch(live_patch, html, streams, events)]);
       }
     }
     attachTrueDocEl() {
@@ -2539,9 +2637,9 @@ within:
       });
       dom_default.all(this.el, `[${this.binding(PHX_MOUNTED)}]`, (el) => this.maybeMounted(el));
     }
-    applyJoinPatch(live_patch, html, events) {
+    applyJoinPatch(live_patch, html, streams, events) {
       this.attachTrueDocEl();
-      let patch = new DOMPatch(this, this.el, this.id, html, null);
+      let patch = new DOMPatch(this, this.el, this.id, html, streams, null);
       patch.markPrunableContentForRemoval();
       this.performPatch(patch, false);
       this.joinNewChildren();
@@ -2713,8 +2811,8 @@ within:
         });
       } else if (!isEmpty(diff)) {
         this.liveSocket.time("full patch complete", () => {
-          let html = this.renderContainer(diff, "update");
-          let patch = new DOMPatch(this, this.el, this.id, html, null);
+          let [html, streams] = this.renderContainer(diff, "update");
+          let patch = new DOMPatch(this, this.el, this.id, html, streams, null);
           phxChildrenAdded = this.performPatch(patch, true);
         });
       }
@@ -2727,15 +2825,15 @@ within:
       return this.liveSocket.time(`toString diff (${kind})`, () => {
         let tag = this.el.tagName;
         let cids = diff ? this.rendered.componentCIDs(diff).concat(this.pruningCIDs) : null;
-        let html = this.rendered.toString(cids);
-        return `<${tag}>${html}</${tag}>`;
+        let [html, streams] = this.rendered.toString(cids);
+        return [`<${tag}>${html}</${tag}>`, streams];
       });
     }
     componentPatch(diff, cid) {
       if (isEmpty(diff))
         return false;
-      let html = this.rendered.componentToString(cid);
-      let patch = new DOMPatch(this, this.el, this.id, html, cid);
+      let [html, streams] = this.rendered.componentToString(cid);
+      let patch = new DOMPatch(this, this.el, this.id, html, streams, cid);
       let childrenAdded = this.performPatch(patch, true);
       return childrenAdded;
     }
@@ -3614,7 +3712,7 @@ within:
       return rootsFound;
     }
     redirect(to, flash) {
-      this.disconnect();
+      this.unload();
       browser_default.redirect(to, flash);
     }
     replaceMain(href, flash, callback = null, linkRef = this.setPendingLink(href)) {
@@ -4124,7 +4222,6 @@ within:
     constructor() {
       this.transitions = new Set();
       this.pendingOps = [];
-      this.reset();
     }
     reset() {
       this.transitions.forEach((timer) => {
@@ -4145,9 +4242,7 @@ within:
       let timer = setTimeout(() => {
         this.transitions.delete(timer);
         onDone();
-        if (this.size() === 0) {
-          this.flushPendingOps();
-        }
+        this.flushPendingOps();
       }, time);
       this.transitions.add(timer);
     }
@@ -4158,8 +4253,14 @@ within:
       return this.transitions.size;
     }
     flushPendingOps() {
-      this.pendingOps.forEach((op) => op());
-      this.pendingOps = [];
+      if (this.size() > 0) {
+        return;
+      }
+      let op = this.pendingOps.shift();
+      if (op) {
+        op();
+        this.flushPendingOps();
+      }
     }
   };
   return phoenix_live_view_exports;